Polished a tiny bit
Now the shadows growth is calculated so that they grow at a speed dependant on the arrows distance from the target.
I.e. The shadows of arrows targeting column one grow quicker.
Right now the max shadow size is 100. Another idea is to have the same growth rate for all shadows but have the shadows start smaller the farther right you are.

Also made it so that the arrows in the middle column hit the shield by changing the 700 destination to 685. Maitiu thought that this looked weird, but having only those arrows offset from the shield looks weirder.